requirements.txt
Django
pytz
pymongo
djongo
djangorestframework
django-cors-headers